move the effect to feature flags instead of segments
